+++ plugins/postcss-gradients-interpolation-method/package.json
@@ -32,8 +32,7 @@
 		"@csstools/css-color-parser": "^0.0.0",
 		"@csstools/css-parser-algorithms": "^2.0.1",
 		"@csstools/css-tokenizer": "^2.1.0",
-		"@csstools/postcss-progressive-custom-properties": "^2.0.0",
-		"postcss-value-parser": "^4.2.0"
+		"@csstools/postcss-progressive-custom-properties": "^2.0.0"
 	},
 	"peerDependencies": {
 		"postcss": "^8.4"
